[feature] Added meta tags to head + changed routes (#99)
